+++ src/Commands/DeselectObject.cpp
@@ -1,0 +1,34 @@
+#pragma once
+#include "DeselectObject.h"
+#include <Utilities/GeneralUtils.h>
+#include <Ext/Techno/Body.h>
+
+const char* DeselectObjectCommandClass::GetName() const
+{
+	return "Deselect 1 object from current selection";
+}
+
+const wchar_t* DeselectObjectCommandClass::GetUIName() const
+{
+	return GeneralUtils::LoadStringUnlessMissing("TXT_DESELECT", L"Deselect 1 Object");
+}
+
+const wchar_t* DeselectObjectCommandClass::GetUICategory() const
+{
+	return CATEGORY_SELECTION;
+}
+
+const wchar_t* DeselectObjectCommandClass::GetUIDescription() const
+{
+	return GeneralUtils::LoadStringUnlessMissing("TXT_DESELECT_DESC", L"Deselect 1 object from current selection.");
+}
+
+void DeselectObjectCommandClass::Execute(WWKey eInput) const
+{
+	int nCount = ObjectClass::CurrentObjects.Count;
+
+	if (nCount > 0)
+	{
+		ObjectClass::CurrentObjects.GetItem(0)->Deselect();
+	}
+}